Refactor user repository and implementation
