+++ internal/repository/userrepo/user_repo_impl.go
@@ -4,36 +4,40 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/bajra-manandhar17/personal-finance-app/internal/model"
-	"gorm.io/gorm"
+	"github.com/bajra-manandhar17/personal-finance-app/internal/db/model"
+	"github.com/bajra-manandhar17/personal-finance-app/internal/db/query"
 )
 
 type UserRepoImpl struct {
-	db *gorm.DB
+	query *query.Query
 }
 
-func (u *UserRepoImpl) CreateUser(ctx context.Context, user *model.User) error {
-	if err := u.db.Create(user).Error; err != nil {
+func (u *UserRepoImpl) CreateUser(ctx context.Context, user *model.Users) error {
+	if err := u.newUserQuery(ctx).Create(user); err != nil {
 		return fmt.Errorf("error creating user: %w", err)
 	}
 
 	return nil
 }
 
-func (u *UserRepoImpl) GetUser(ctx context.Context, userId string) (*model.User, error) {
-	var user model.User
-
-	if err := u.db.Where("user_id = ?", userId).First(&user).Error; err != nil {
+func (u *UserRepoImpl) GetUser(ctx context.Context, userId string) (*model.Users, error) {
+	user, err := u.newUserQuery(ctx).Where(u.query.Users.UserID.Eq(userId)).First()
+	if err != nil {
 		return nil, fmt.Errorf("error getting user: %w", err)
 	}
-	return &user, nil
+
+	return user, nil
 }
 
 func (u *UserRepoImpl) DoesEmailExist(ctx context.Context, email string) (bool, error) {
-	var count int64
-	if err := u.db.Model(&model.User{}).Where("email = ?", email).Count(&count).Error; err != nil {
+	count, err := u.newUserQuery(ctx).Where(u.query.Users.Email.Eq(email)).Count()
+	if err != nil {
 		return false, fmt.Errorf("error checking if email exists: %w", err)
 	}
 
 	return count > 0, nil
+}
+
+func (u *UserRepoImpl) newUserQuery(ctx context.Context) query.IUsersDo {
+	return u.query.Users.WithContext(ctx)
 }
